remove Launcher.rank_id
