+++ legate/driver/command.py
@@ -27,6 +27,10 @@ if TYPE_CHECKING:
 __all__ = ("CMD_PARTS",)
 
 
+# this will be replaced by bind.sh with the actual computed rank at runtime
+LEGATE_GLOBAL_RANK_SUBSTITUTION = "%%LEGATE_GLOBAL_RANK%%"
+
+
 def cmd_bind(
     config: ConfigProtocol, system: System, launcher: Launcher
 ) -> CommandPart:
@@ -100,7 +104,10 @@ def cmd_nvprof(
     if not config.profiling.nvprof:
         return ()
 
-    log_path = str(config.logging.logdir / f"legate_{launcher.rank_id}.nvvp")
+    log_path = str(
+        config.logging.logdir
+        / f"legate_{LEGATE_GLOBAL_RANK_SUBSTITUTION}.nvvp"
+    )
 
     return ("nvprof", "-o", log_path)
 
@@ -111,7 +118,9 @@ def cmd_nsys(
     if not config.profiling.nsys:
         return ()
 
-    log_path = str(config.logging.logdir / f"legate_{launcher.rank_id}")
+    log_path = str(
+        config.logging.logdir / f"legate_{LEGATE_GLOBAL_RANK_SUBSTITUTION}"
+    )
     targets = config.profiling.nsys_targets
     extra = config.profiling.nsys_extra
